prescient/pyproject.toml
[project]
name = "prescient"
version = "0.2.0"
description = "Autonomous Prediction Markets Powered by AI Agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Julian Ramirez", email = "juramirez.ve@gmail.com"},
]
keywords = ["ai", "agent", "prediction-market", "defi", "uniswap", "dune", "filecoin"]

dependencies = [
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "web3>=6.15.0",
    "pandas>=2.2.0",
    "httpx>=0.27.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "eth-account>=0.11.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
prescient-agent = "agent.orchestrator:cli_main"
prescient-api = "api.main:cli_main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agent*", "api*"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
